requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
PySocks>=1.7.1